backend/requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
pdfplumber==0.11.7
pydantic==2.12.3
python-multipart==0.0.22
python-dotenv==1.2.1